requirements.txt
python-dotenv
beautifulsoup4
google-generativeai
requests 
